Retune some kernels.
Retune radix sort, run length encoding, reduce by key, scan, select if, and
histogram for SM70 and SM80.

+++ test/test_device_reduce_by_key.cu
@@ -488,7 +488,7 @@ void Test(
     {
         float   avg_millis  = elapsed_millis / g_timing_iterations;
         float   giga_rate   = float(num_items) / avg_millis / 1000.0f / 1000.0f;
-        int     bytes_moved = ((num_items + num_segments) * sizeof(KeyT)) + ((num_items + num_segments) * sizeof(ValueT));
+        size_t  bytes_moved = (size_t(num_items + num_segments) * sizeof(KeyT)) + ((size_t(num_items) + num_segments) * sizeof(ValueT));
         float   giga_bandwidth  = float(bytes_moved) / avg_millis / 1000.0f / 1000.0f;
         printf(", %.3f avg ms, %.3f billion items/s, %.3f logical GB/s", avg_millis, giga_rate, giga_bandwidth);
     }
